alarm detection more realistic
The alarm detection is kitchen_node has been completely changed. Now the code is more similar to  the one to be actually used in rela sensors. E.g., che temperature is actually periodically sampled, only a trick is used in order to improve its value when a button is press.